reorganize secrets plugins
