+++ app/main.go
@@ -14,6 +14,7 @@ import (
 	"github.com/winhowes/AuthTransformer/app/authplugins"
 	_ "github.com/winhowes/AuthTransformer/app/authplugins/incoming"
 	_ "github.com/winhowes/AuthTransformer/app/authplugins/outgoing"
+	_ "github.com/winhowes/AuthTransformer/app/secrets/plugins"
 )
 
 type Config struct {
