+++ app/secrets/plugins/env.go
@@ -1,6 +1,10 @@
-package secrets
+package plugins
 
-import "os"
+import (
+	"os"
+
+	"github.com/winhowes/AuthTransformer/app/secrets"
+)
 
 // envPlugin loads secrets from environment variables.
 type envPlugin struct{}
@@ -9,4 +13,4 @@ func (envPlugin) Prefix() string { return "env" }
 
 func (envPlugin) Load(id string) (string, error) { return os.Getenv(id), nil }
 
-func init() { Register(envPlugin{}) }
+func init() { secrets.Register(envPlugin{}) }
